Fix ensuring directory exist in multi thread bug.
